// src/app/shared/enum/phau-thuat-thu-thuat.enum.ts
// export enum EnumTrangThaiPhauThuatThuThuat
// {
//     DangPhauThuat = 1,
//     TheoDoi = 2,
//     ChuyenGiaoBenhNhan = 3,
//     TuVong = 4
// }

export enum EnumVaiTroBacSi
{
    PhauThuatVienChinh = 1,
    GayMeTeChinh = 2,
    GayMeTePhu = 3,
    BacSiHoiChan = 4,
    BacSiTangCuong = 5,
    Phu1 = 6,
    Phu2 = 7,
    Phu3 = 8
}

export enum EnumVaiTroDieuDuong
{
    PhuPhauThuat = 1,
    DungCuVongTrong = 2,
    DungCuVongNgoai = 3,
    Phu1 = 4,
    Phu2 = 5,
    Phu3 = 6,
    PhuMePhuTe = 7,
    ChayNgoai = 8,
    PhauThuatVienChinh = 9
}

export enum EnumTrangThaiPhauThuatThuThuat
{
    ChoPhauThuat = 1,
    DangPhauThuat = 2,
    TheoDoi = 3,
    ChuyenGiao = 4,
    TuVong = 5,
    KhongCo = 6
}

export enum EnumTrangThaiTheoDoiSauPhauThuatThuThuat
{
    DangTheoDoi = 1,
    KetThucTheoDoi = 2
}

export enum EnumGiaiDoanPhauThuat
{
    GayMe = 1,
    PhauThuat = 2,
    HoiTinh = 3,
    Khac = 4
}

export enum EnumLoaiNoiChiDinh
{
    NoiTruPhieuDieuTri = 1,
    PhauThuatThuThuat = 2,
    ChanDoanHinhAnh = 3,
    YeuCauKhamBenh = 4,
    TiemChung = 5   
}